fix : move rest local dtos to repository module
