+++ repository/src/main/java/com/cairosquad/repository/search/data_source/local/dto/RecentSearchEntity.kt
@@ -1,4 +1,4 @@
-package com.cairosquad.local.search.recent.entity
+package com.cairosquad.repository.search.data_source.local.dto
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
